Allow specifying clang version in makefile
This reverts the CLANG specific checks added in
9eb2ff504ac25dc9d7f95ebc59660fd1ec289174, because they would break if a
specific CLANG version was specified instead of a generic one.

I hope to be able to make it so that you don't use CLANG even if your
make command has CLANG=0, but as these changes are they break things, so
I'll revert them and work on that later.